Adds broadcast view back to broadcast views file - to reduce diff
