+++ app/main/views/broadcast_common.py
@@ -1,7 +1,5 @@
-from flask import redirect, request, url_for
+from flask import request
 
-from app import current_service
-from app.broadcast_areas.models import CustomBroadcastAreas
 from app.main import main
 from app.main.views.broadcast import (
     choose_broadcast_area,
@@ -26,11 +24,8 @@ from app.main.views.templates import (
     search_postcodes_for_template,
     write_new_broadcast_from_template,
 )
-from app.models.broadcast_message import BroadcastMessage
-from app.models.template import Template
 from app.utils import service_has_permission
-from app.utils.broadcast import redirect_dependent_on_alert_area
-from app.utils.user import user_has_any_permissions, user_has_permissions
+from app.utils.user import user_has_any_permissions
 
 
 @main.route("/services/<uuid:service_id>/write-new-broadcast", methods=["GET", "POST"])
@@ -50,66 +45,6 @@ def write_new_broadcast(service_id):
         return write_new_broadcast_from_template(service_id=service_id, template_id=template_id)
     else:
         return create_new_broadcast(service_id)
-
-
-@main.route("/services/<uuid:service_id>/new-broadcast/<uuid:template_id>")
-@user_has_permissions("create_broadcasts", restrict_admin_usage=True)
-@service_has_permission("broadcast")
-@user_has_any_permissions(["create_broadcasts", "manage_templates"], restrict_admin_usage=True)
-def broadcast(service_id, template_id):
-    template = Template.from_id(template_id=template_id, service_id=service_id)
-    broadcast_message = None
-    if template.reference and template.content:
-        if template.areas:
-            # As Template area already exists, created broadcast_message using this
-            # and reference and content if they have been set also
-            broadcast_message = (
-                BroadcastMessage.create_from_custom_area(
-                    service_id=service_id,
-                    template_id=template_id,
-                    areas=template.areas,
-                    content=template.content,
-                    reference=template.reference,
-                )
-                if type(template.areas) is CustomBroadcastAreas
-                else BroadcastMessage.create_from_area(
-                    service_id=service_id,
-                    template_id=template_id,
-                    area_ids=template.area_ids,
-                    content=template.content,
-                    reference=template.reference,
-                )
-            )
-        else:
-            # Only reference and content have been set for Template,
-            # so can only use these to create broadcast_message
-            broadcast_message = BroadcastMessage.create_from_content(
-                service_id=service_id, content=template.content, reference=template.reference
-            )
-
-        # If the current service is an operator service, user is redirected
-        # straight to choose area as extra_content attribute shouldn't be set for
-        # alerts created in operator services
-        if current_service.broadcast_channel == "operator":
-            return redirect_dependent_on_alert_area(broadcast_message)
-        else:
-            return redirect(
-                url_for(
-                    ".choose_extra_content",
-                    service_id=current_service.id,
-                    broadcast_message_id=broadcast_message.id,
-                )
-            )
-    elif template.areas:
-        # ONLY Template area has been set, reference and content must be provided to
-        # create broadcast_message so redirected to relevant page
-        return redirect(
-            url_for(
-                ".write_new_broadcast",
-                service_id=current_service.id,
-                template_id=template.id,
-            )
-        )
 
 
 @main.route(
